refactor to incorporate data generation inside server/index.js
