+++ spec/serverSpec.js
@@ -3,7 +3,7 @@ const Promise = require('bluebird');
 const { Pool } = require('pg');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const config = require('../database/config');
+const config = require('../credentials/config');
 const db = require('../database/database');
 const elastic = require('../dashboard/elastic');
 const setup = require('../database/setup');
